requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
scikit-learn
numpy